requirements.txt
streamlit
gspread
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pandas
streamlit-oauth